refactor: User도메인의 Controller와 Service 클래스 패키지로 버전을 구분
