+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/auth/controller/AuthController.java
@@ -6,7 +6,7 @@ import com.hertz.hertz_be.domain.auth.responsecode.AuthResponseCode;
 import com.hertz.hertz_be.domain.auth.repository.RefreshTokenRepository;
 import com.hertz.hertz_be.domain.auth.service.AuthService;
 import com.hertz.hertz_be.domain.auth.dto.request.TestLoginRequestDto;
-import com.hertz.hertz_be.domain.user.service.UserService;
+import com.hertz.hertz_be.domain.user.service.v1.UserService;
 import com.hertz.hertz_be.global.auth.token.JwtTokenProvider;
 import com.hertz.hertz_be.global.common.ResponseDto;
 import com.hertz.hertz_be.global.exception.BusinessException;
